+++ tests/testthat/test-getSimulatedData.R
@@ -1,4 +1,5 @@
 # getOneToTwoRows ----
+
 test_that("getOneToTwoRows works as expected", {
   simDataOne <- data.frame(
     id = c(1:3), to = c(1, 1, 1), from = c(0, 0, 0), entry = c(0, 0, 0),
@@ -17,7 +18,8 @@ test_that("getOneToTwoRows works as expected", {
   expect_true(all(actualCens$censTime == actualCens$exit))
 })
 
-# getSimulatedData ---
+# getSimulatedData ----
+
 test_that("getSimulatedData works as expected if no random censoring is present", {
   actual <- getSimulatedData(10,
     transition = exponential_transition(h01 = 1, h02 = 1.5, h12 = 1),
@@ -55,8 +57,6 @@ test_that("getSimulatedData works as expected if random censoring is present", {
   expect_true(all(actual$from %in% c(0, 1)))
 })
 
-
-
 test_that("getSimulatedData creates expected data set for exponential transitions", {
   set.seed(1243)
   actual <- getSimulatedData(4,
@@ -89,4 +89,24 @@ test_that("getSimulatedData creates expected data set for Weibull transitions", 
     stringsAsFactors = FALSE
   )
   expect_equal(actual[1, ], row1)
+})
+
+test_that("getSimulatedData works also without progression to death transitions", {
+  set.seed(31)
+  result <- expect_silent(getSimulatedData(
+    N = 1L,
+    transition = piecewise_exponential(
+      pw01 = c(0, 2),
+      pw02 = c(0, 2),
+      pw12 = c(0, 2),
+      h01 = rep(0.04781994, 2),
+      h02 = rep(0.03882345, 2),
+      h12 = rep(0.31313900, 2)
+    ),
+    dropout = list(rate = 0, time = 1),
+    accrual = list(param = "intensity", value = 5)
+  ))
+  expect_data_frame(result, nrows = 1L)
+  expect_false(any(result$to == 1))
+  expect_false(any(result$from == 1))
 })